+++ shepherd_core/shepherd_core/vsource/virtual_source_simulation.py
@@ -11,12 +11,14 @@ from contextlib import ExitStack
 from pathlib import Path
 from typing import Optional
 
+import numpy as np
 from tqdm import tqdm
 
 from .. import CalibrationEmulator
 from .. import Reader
 from .. import Writer
 from ..data_models import VirtualSourceConfig
+from ..logger import logger
 from . import VirtualSourceModel
 from .target_model import TargetABC
 
@@ -26,6 +28,8 @@ def simulate_source(
     target: TargetABC,
     path_input: Path,
     path_output: Optional[Path] = None,
+    *,
+    monitor_internals: bool = False,
 ) -> float:
     """Simulate behavior of virtual source algorithms.
 
@@ -47,10 +51,25 @@ def simulate_source(
         cal_out = file_out.get_calibration_data()
 
     src = VirtualSourceModel(
-        config, cal_emu, log_intermediate=False, window_size=file_inp.get_window_samples()
+        config,
+        cal_emu,
+        dtype_in=file_inp.get_datatype(),
+        log_intermediate=False,
+        window_size=file_inp.get_window_samples(),
     )
     i_out_nA = 0
     e_out_Ws = 0.0
+    if monitor_internals and path_output:
+        stats_sample = 0
+        stats_internal = np.empty((round(file_inp.runtime_s * file_inp.samplerate_sps), 11))
+        try:
+            # keep dependencies low
+            from matplotlib import pyplot as plt
+        except ImportError:
+            logger.warning("Matplotlib not installed, plotting of internals disabled")
+            stats_internal = None
+    else:
+        stats_internal = None
 
     for _t, v_inp, i_inp in tqdm(
         file_inp.read_buffers(is_raw=True), total=file_inp.buffers_n, desc="Buffers", leave=False
@@ -66,7 +85,22 @@ def simulate_source(
             )
             i_out_nA = target.step(int(v_uV[_n]), pwr_good=src.cnv.get_power_good())
             i_nA[_n] = i_out_nA
-            # TODO: src.cnv.get_I_mod_out_nA() has more internal drains
+
+            if stats_internal is not None:
+                stats_internal[stats_sample] = [
+                    _t[_n] * 1e-9,  # s
+                    src.hrv.voltage_hold * 1e-6,
+                    src.cnv.V_input_request_uV * 1e-6,  # V
+                    src.hrv.voltage_set_uV * 1e-6,
+                    src.cnv.V_mid_uV * 1e-6,
+                    src.hrv.current_hold * 1e-6,  # mA
+                    src.hrv.current_delta * 1e-6,
+                    i_out_nA * 1e-6,
+                    src.cnv.P_inp_fW * 1e-12,  # mW
+                    src.cnv.P_out_fW * 1e-12,
+                    src.cnv.get_power_good(),
+                ]
+                stats_sample += 1
 
         e_out_Ws += (v_uV * i_nA).sum() * 1e-15 * file_inp.sample_interval_s
         if path_output:
@@ -75,4 +109,36 @@ def simulate_source(
             file_out.append_iv_data_raw(_t, v_out, i_out)
 
     stack.close()
+
+    if stats_internal is not None:
+        stats_internal = stats_internal[:stats_sample, :]
+        fig, axs = plt.subplots(4, 1, sharex="all", figsize=(20, 4 * 6), layout="tight")
+        fig.suptitle(f"VSrc-Sim with {config.name}, Inp={path_input.name}, E={e_out_Ws} Ws")
+        axs[0].set_ylabel("Voltages [V]")
+        axs[0].plot(stats_internal[:, 0], stats_internal[:, 1:5])
+        axs[0].legend(["V_cv_hold", "V_inp_Req", "V_cv_set", "V_cap"], loc="upper right")
+
+        axs[1].set_ylabel("Current [mA]")
+        axs[1].plot(stats_internal[:, 0], stats_internal[:, 5:8])
+        axs[1].legend(["C_cv_hold", "C_cv_delta", "C_out"], loc="upper right")
+
+        axs[2].set_ylabel("Power [mW]")
+        axs[2].plot(stats_internal[:, 0:1], stats_internal[:, 8:10])
+        axs[2].legend(["P_inp", "P_out"], loc="upper right")
+
+        axs[3].set_ylabel("PwrGood [n]")
+        axs[3].plot(stats_internal[:, 0], stats_internal[:, 10])
+        axs[3].legend(["PwrGood"], loc="upper right")
+
+        axs[3].set_xlabel("Runtime [s]")
+
+        for ax in axs:
+            # deactivates offset-creation for ax-ticks
+            ax.get_yaxis().get_major_formatter().set_useOffset(False)
+            ax.get_xaxis().get_major_formatter().set_useOffset(False)
+
+        plt.savefig(path_output.with_suffix(".png"))
+        plt.close(fig)
+        plt.clf()
+
     return e_out_Ws