+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # History of Changes
 
+## v2024.9.1
+
+- virtual harvester
+  - allow direct pass-through
+  - cv-harvester can now extrapolate (linear around current set-point) which makes it more responsive and reduces error
+- virtual source
+  - add feedback-path to harvester (to control cv-harvester) when no boost-converter is used
+  - bring changes from pru-code to py
+  - simulation can now plot several internal states
+  - simulation example now runs several models through the jogging-dataset and creates plots and data
+- virtual targets
+  - improve current targets
+  - add diode+resistor target (to emulate a diode for burning energy)
+  - add exemplary instantiations of some targets
+- shp-reader - improve calculation of file-stats
+
 ## v2024.8.2
 
 - add hdf5-file based simulations
@@ -222,6 +238,8 @@ sudo python3 -X importtime -c 'from shepherd_core.data_models.task import Emulat
 # 12.8 s on v2024.4.1, pydantic 2.7.0, core 2.18.1
 # 10.3 s on v2024.5.1, pydantic 2.7.4, core 2.18.4 - debian 12.5
 # 10.4 s on v2024.5.1, pydantic 2.8.0, core 2.20.0
+# 12.3 s on v2024.8.2, pydantic 2.8.2, core 2.20.1
+# 11.7 s on v2024.8.2, pydantic 2.9.0, core 2.23.2
 ```
 
 ## v2023.08.6